src/node_c/requirements.txt
flask==3.0.0
gunicorn==21.2.0
google-cloud-storage>=2.14.0
diffusers>=0.30.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
torch>=2.1.0
rembg>=2.0.0
pillow>=10.0.0
